==> 2_prediction.ipynb <==
import tensorflow as tf
from tensorflow.keras.models import load_model
import pandas as pd
import numpy as np
import pickle

# load the trained model
model = load_model('artifacts/classification/classification_model.keras')

# load the scaler
with open('artifacts/classification/scaler.pkl', 'rb') as f:
    scaler = pickle.load(f)

# load the encoders
with open('artifacts/classification/label_encoder_gender.pkl', 'rb') as f:
    label_encoder_gender = pickle.load(f)

with open('artifacts/classification/onehot_encoder_geo.pkl', 'rb') as f:
    onehot_encoder_geo = pickle.load(f)

# Example input data
input_data = {
    'CreditScore': 600,
    'Geography': 'France',
    'Gender': 'Male',
    'Age': 40,
    'Tenure': 3,
    'Balance': 60000,
    'NumOfProducts': 2,
    'HasCrCard': 1,
    'IsActiveMember': 1,
    'EstimatedSalary': 50000
}

# one hot encode 'Geography'
geo_encoded = onehot_encoder_geo.transform([[input_data['Geography']]])
geo_df = pd.DataFrame(geo_encoded, columns=onehot_encoder_geo.get_feature_names_out(['Geography']))
geo_df

input_df = pd.DataFrame([input_data])
input_df

# encode 'Gender'
input_df['Gender'] = label_encoder_gender.transform(input_df['Gender'])
input_df

# concate the one hot encoded geography columns
input_df = pd.concat([input_df.drop('Geography', axis=1), geo_df], axis=1)
input_df

# scale the input data
input_scaled = scaler.transform(input_df)
input_scaled

# predict churn probability
prediction = model.predict(input_scaled)
prediction

churn_probability = prediction[0][0]

churn_probability

if churn_probability > 0.5:
    print(f'The customer is likely to churn')
else:
    print(f'The customer is unlikely to churn')
